+++ src/components/Menu/MenuDivider.tsx
@@ -1,9 +1,26 @@
 import React from 'react';
-import { MenuDivider as BlueprintMenuDivider } from "@blueprintjs/core";
+import {
+    MenuDivider as BlueprintMenuDivider,
+    MenuDividerProps as BlueprintMenuDividerProps,
+} from "@blueprintjs/core";
 import { CLASSPREFIX as eccgui } from "../../configuration/constants";
 
-function MenuDivider({children, className='', ...restProps}: any) {
-    return <BlueprintMenuDivider {...restProps} className={`${eccgui}-menu__divider ` + className}>{children}</BlueprintMenuDivider>;
+export interface MenuDividerProps extends BlueprintMenuDividerProps {
+    // we do not add own changes here currently
+}
+
+function MenuDivider({
+    children,
+    className="",
+    ...restProps
+}: MenuDividerProps) {
+    return (
+        <BlueprintMenuDivider
+            {...restProps}
+            className={`${eccgui}-menu__divider ` + className}>
+            {children}
+        </BlueprintMenuDivider>
+    );
 }
 
 export default MenuDivider;